[zen-12] Create factory-backed form implementation
- Moved base form implementation to BaseForm
- Added "create()" method to Factory; allows creating elements,
  fieldsets, or forms, based on "type" parameter provided
- Created FormFactoryAwareInterface, for composing form factories
- Created new "Form" implementation that is factory-backed
  - implements FormFactoryAwareInterface
  - overrides add() to create elements/fieldsets from specs before
    attaching
